+++ src/briefy/leica/views/assignments.py
@@ -44,7 +44,7 @@ class AssignmentService(RESTService):
     friendly_name = model.__name__
     default_order_by = 'updated_at'
     filter_related_fields = [
-        'project.title', 'title', 'professional.title', 'professional.main_email',
+        'project.title', 'title', 'professional.title', 'professional.email',
         'project.id', 'description', '_location.locality', '_location.country',
         'location.formatted_address', 'location.fullname', 'location.email',
         'professional_user', 'project_manager', 'scout_manager', 'qa_manager',